Done & approval from 4th meeting

+++ Source/Component/SurfaceComponent.cpp
@@ -2,7 +2,7 @@
 
 #include <glm/glm.hpp>
 
-SurfaceComponent::SurfaceComponent(const glm::vec3& a, const glm::vec3& b, const glm::vec3& c, const sf::Color& color):
+SurfaceComponent::SurfaceComponent(const glm::dvec3& a, const glm::dvec3& b, const glm::dvec3& c, const sf::Color& color):
 	m_Vertices{a, b, c },
 	m_Color(color)
 {
