+++ Source/Component/SurfaceComponent.hpp
@@ -9,11 +9,11 @@ class SurfaceComponent
 {
 public:
 	// Because the surface is a triangle
-	std::array<glm::vec3, 3> m_Vertices;
+	std::array<glm::dvec3, 3> m_Vertices;
 	sf::Color m_Color;
 
 	SurfaceComponent() = default;
-	SurfaceComponent(const glm::vec3& a, const glm::vec3& b, const glm::vec3& c, const sf::Color& color = sf::Color::Green);
+	SurfaceComponent(const glm::dvec3& a, const glm::dvec3& b, const glm::dvec3& c, const sf::Color& color = sf::Color::Green);
 	~SurfaceComponent() = default;
 };
