+++ Source/System/PainterRenderSystem.hpp
@@ -5,17 +5,12 @@
 #include <tuple>
 #include <entt/entt.hpp>
 #include <SFML/Graphics.hpp>
-#include <graphviz/gvc.h>
-#include <graphviz/cgraph.h>
-#include <graphviz/gvplugin.h>
 
 #include "../Macro.hpp"
 #include "../Component/SurfaceComponent.hpp"
 #include "../Component/Surface2DComponent.hpp"
 #include "../Misc/EdgeBucket.hpp"
 #include "../Misc/BinaryTreePartitioning.hpp"
-
-//extern gvplugin_library_t gvplugin_dot_layout_LTX_library;
 
 // Forward declaration
 class BaseRenderSystem;
@@ -38,7 +33,7 @@ public:
 
 	void PrintBinaryPartitioningTree();
 
-	std::tuple<EdgeBucket2, int> GetEdgeBucket(glm::vec3& a, glm::vec3& b);
+	std::tuple<EdgeBucket2, int> GetEdgeBucket(glm::dvec3& a, glm::dvec3& b);
 };
 
 #endif
