+++ Source/System/TransformSystem.cpp
@@ -8,30 +8,30 @@
 
 #define PI 3.14159265358979323846f
 
-static inline float Deg2Rad(float deg)
+static inline double Deg2Rad(double deg)
 {
-	return deg * PI / 180.0f;
+	return deg * PI / 180.0;
 }
 
 void TransformSystem::Reset(entt::registry& registry, entt::entity& entity)
 {
 	TransformComponent& transformComponent = registry.get<TransformComponent>(entity);
-	transformComponent.m_MatTransform = glm::mat4(1.0f);
-	transformComponent.m_MatRotation = glm::mat4(1.0f);
-	transformComponent.m_MatTranslation = glm::mat4(1.0f);
+	transformComponent.m_MatTransform = glm::dmat4(1.0);
+	transformComponent.m_MatRotation = glm::dmat4(1.0);
+	transformComponent.m_MatTranslation = glm::dmat4(1.0);
 }
 
-void TransformSystem::Rotate(entt::registry& registry, entt::entity& entity, glm::vec3 const& rotateVec, float rotationDegree)
+void TransformSystem::Rotate(entt::registry& registry, entt::entity& entity, glm::dvec3 const& rotateVec, double rotationDegree)
 {
 	TransformComponent& transformComponent = registry.get<TransformComponent>(entity);
-	transformComponent.m_MatRotation *= glm::rotate(glm::mat4(1.0f), Deg2Rad(rotationDegree), rotateVec);
+	transformComponent.m_MatRotation *= glm::rotate(glm::dmat4(1.0), Deg2Rad(rotationDegree), rotateVec);
 	UpdateMatTransform(registry, entity);
 }
 
-void TransformSystem::Translate(entt::registry& registry, entt::entity& entity, glm::vec3 const& translateVec)
+void TransformSystem::Translate(entt::registry& registry, entt::entity& entity, glm::dvec3 const& translateVec)
 {
 	TransformComponent& transformComponent = registry.get<TransformComponent>(entity);
-	transformComponent.m_MatTranslation *= glm::translate(glm::mat4(1.0f), translateVec);
+	transformComponent.m_MatTranslation *= glm::translate(glm::dmat4(1.0), translateVec);
 	UpdateMatTransform(registry, entity);
 }
