+++ Source/System/ZBufferRenderSystem.hpp
@@ -19,7 +19,7 @@ class BaseRenderSystem;
 class ZBufferRenderSystem
 {
 public:
-	std::array<std::array<float, WINDOW_WIDTH>, WINDOW_HEIGHT> m_ZBuffer;
+	std::array<std::array<double, WINDOW_WIDTH>, WINDOW_HEIGHT> m_ZBuffer;
 	std::vector<EdgeBucket> m_ActiveEdges;
 
 	ZBufferRenderSystem();
@@ -31,7 +31,7 @@ public:
 
 	void ResetZBuffer();
 
-	std::tuple<EdgeBucket, int> GetEdgeBucket(glm::vec3& a, glm::vec3& b);
+	std::tuple<EdgeBucket, int> GetEdgeBucket(glm::dvec3& a, glm::dvec3& b);
 };
 
 #endif
